Set smooth scroll offset for nav-bar, body class during scrolling.
The current method for setting offset (adding class .js-bolt-smooth-scroll-offset to the fixed element) doesn't work for sticky elements. It uses the element's location on the page instead of the element's height when fixed. With a sticky element halfway down the page, it sets the offset to halfway down the page.

This change sets a class on the body while scrolling, which will allow us to keep the header collapsed during scrolling. Right now, using the headroom.js, the header is collapsed when scrolling down, but opens when scrolling up, which messes up the offset. Instead of offsetting for the height of the nav-bar, it'd need to be the combined height of the header and the navbar. Rather than getting into dynamic offsets for different scenarios, it's easier (and probably more user-friendly) to keep the header collapsed using a scrolling-specific class.