refactor: Rely fully on attribute mirroring instead for nav bar (BK-157)
Previously, the activateLink method was used to do some of the
activation tasks while attribute mirroring was used for others.  This
consolidates the activation heavy lifting in these three methods:

get active()
set active()
attributeChangedCallback()

A major benefit is that setting the 'active' attribute externally now
has the same affect as the previous activateLink method.