Add gumshoe functionality to navbar (BK-157 ) (#396)
* feat: Add initial gumshoe functionality (BK-157)

* refactor: Rely fully on attribute mirroring instead for nav bar (BK-157)

Previously, the activateLink method was used to do some of the
activation tasks while attribute mirroring was used for others.  This
consolidates the activation heavy lifting in these three methods:

get active()
set active()
attributeChangedCallback()

A major benefit is that setting the 'active' attribute externally now
has the same affect as the previous activateLink method.

* fix: Upgrade the 'active' property for nav-link (BK-157)

This resolves a bug where if a given link was initially set to active
because of the page hash, links that appeared after that link in the DOM
would not activate properly when clicked.  The problem was that the
upgraded.

* feat: Initialize gumshoe from within nav-list (BK-157)

This commit does a few things:
- Initializes gumshoe from within the nav-list component, so if there's
no nav-list, nothing happens.  We take care to initialize it only once
even if there are multiple nav-lists on a given page
- Allows consumers to specify the offset in pixels as a property of
nav-list. That way, consumers can adjust for their own fixed headers,
whether by setting the property initially or even updating it dynamically
with their own javascript (if, say, the required offset needs to change
as the header itself changes height dynamically).

* chore: updating package-lock.json

* chore: merging in latest from release/0.x branch