feat: Update pagination schema to move more logic to back-end (BDS-5)
Whereas previously Bolt was only taking parameters for "current" and
"total" number of pages and handling all the logic for how many items
should be shown and their respective URLs, this commit moves
responsibility for that logic to the back end.  We know Drupal handles
this itself, so there will be less friction there; other back-end
systems should be able to handle this logic themselves as well.