feat: Remove all vertical padding on video share overlay (BDS-495)
The share overlay content is vertically centered anyway, so this only
affects cases where the content is slightly too large to fit vertically.
In those cases, the share content now presses closer to the top and
bottom of the video player without activating the scrollbar.  However,
the scrollbar still activates if the content would actually be cropped.