docs: Remove unused "align" property from card examples
When passing buttons to the card component as part of contentItems,
specifying either "align" or "style" property does nothing.  Both of
these are forced to preset values in card.twig

Note that if we wanted to allow this behavior, we could change this:
`{% set item = item | merge(paramOverrides) %}`

To this:
`{% set item = paramOverrides | merge(item) %}`

But, this could have unintended consequences for anyone who's currently
setting those parameters without realizing they aren't being used.
Better to leave it as-is, which may in fact have been intentional,
than to potentially break things.