BDS-569 Remove unused target attribute and urls
Why is the target blank?  The link default behavior is supposed to be
disabled, so it should make no difference.  From some experimentation,
I found that removing it does make a difference in pattern lab, in much
the same way as smooth scroll links don't work (i.e. if works if you
open it in a new window, but not in the pattern lab interface).  I'm
assuming that was the only reason for it.

Why is "url" (which is supposed to be the value of text to copy)
actually set the link's URL?  Again, these links should never actually
work, unless say javascript is disabled.  Is this a sensible javascript
fallback?  It doesn't seem like it to me.