fix: Correct icon color rules
While it seems to work in most browsers (which is to say, nothing breaks),
the line removed in this commit:

1) is unnecessary, because the same rule is repeated withing the SVGs
2) breaks in earlier versions of Edge, e.g. Edge 18.  My best guess is
that that's because `color: currentColor` is circular/redundant.
`var(--bolt-theme-icon)` also appears to work in all browsers, but why
not just remove it if it's not doing anything the SVG itself isn't also
doing?