DS-489: Fix .class syntax
Dot notation does some fancy magic behind the scenes.  Read more here:
https://twig.symfony.com/doc/2.x/templates.html

Something that was added in Twig 2.x, when using foo.bar, it will check
for a `hasBar` method on the underlying object if `bar` is not a property
of it.  In this case, `hasClass` is a method, but not one we intend to
ever call.  This syntax makes that clear to twig, I think.